No luck trying to fix error in lexical analyzer. Will make a way for the syntax analyzer to negate it in the symbol table or something.
