+++ ArnoldC-interpreter/.vs/ArnoldC-interpreter/xs/UserPrefs.xml
@@ -1,7 +1,8 @@
 ﻿<Properties StartupConfiguration="{77675913-E32F-4642-B420-9C558642C79C}|Default">
-  <MonoDevelop.Ide.Workbench ActiveDocument="ArnoldC-interpreter/MainWindow.cs">
+  <MonoDevelop.Ide.Workbench ActiveDocument="ArnoldC-interpreter/Program.cs">
     <Files>
-      <File FileName="ArnoldC-interpreter/MainWindow.cs" Line="1" Column="1" />
+      <File FileName="ArnoldC-interpreter/Lexer.cs" Line="1" Column="1" />
+      <File FileName="ArnoldC-interpreter/Program.cs" Line="21" Column="67" />
     </Files>
   </MonoDevelop.Ide.Workbench>
   <MonoDevelop.Ide.Workspace ActiveConfiguration="Debug|x86" />
