+++ arnoldc samples/lexemes.arnoldc
@@ -1,5 +1,6 @@
 IT'S SHOWTIME
 
-TALK TO THE HAND "hi bitch"
+HEY CHRISTMAS TREE H
+YOU SET US UP 1
 
 YOU HAVE BEEN TERMINATED
